+++ src/lib/plugin/index.ts
@@ -3,5 +3,6 @@ export {
   importAndLoadTesttimePlugins,
   importAndLoadWorktimePlugins,
 } from './load'
+export * from './manifest'
 export * from './types'
 export * from './utils'
